feat: user flow refactor — intelligent routing and onboarding (#107)
- Home page (/) is now a server component that detects user state and
  routes to WelcomeScreen (first-time users) or GoalScreen (returning)
- Generation form moves to /tailor; middleware protects the route
- WelcomeScreen: 3 onboarding paths — upload resume (extract + save as
  default + redirect to /tailor), AI interview (/interview), or paste
  manually (save to library + redirect to /tailor)
- GoalScreen: 5 goal cards (Tailor Now, Update Base Resume, Add More
  Experience, Prep for Interview, View Applications) with the tailor
  option highlighted as primary
- State 2 banner: amber warning when user has docs but no default resume
- State 3 banner: dismissable blue tip when base resume is 30+ days old
- Skip link: "Skip to generation" sets localStorage flag so subsequent
  visits route directly to /tailor
- Navbar: "Tailor New Resume" → /tailor, "My Documents" → "My Profile"
- Dashboard + resumes "Back" links updated to /tailor
- Tour description updated to reference My Profile
